Reorganize RWTH Provider
