+++ providers/rwth/registrar.go
@@ -1,0 +1,3 @@
+package rwth
+
+// No registrar functionality
